% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA earthquake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a dataframe with raw data from the significant earthquake database. This data can be retrieved by the function \code{get_eq_data()}}
}
\value{
This function returns the NOAA earthquake data with an added date column and coordinates converted to numeric class.
}
\description{
This function adds a date column to the NOAA significant earthquake date and converts longtitude and latitude to numeric
}
\note{
If no month or day is available it is automatically set to 1.
}
\examples{
\dontrun{

clean_data <- eq_clean_data(raw_data)
clean_data <- eq_clean_data(get_eq_data())
}
}
