% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_earthquakes.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create a Label for the Earthquake map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a (cleaned) dataframe with earthquake data from the NOAA earthquake database}
}
\value{
The input dataframe with an additional column called popup_text, which contains (when available) in html the Location,
the Magnitude and the Total number of deaths for each earthquake
}
\description{
This function adds a column \code{popup_text}  to the earthquake dataframe. This column can be used for the annotation column in \code{eq_map()}
}
\examples{
\dontrun{

eq_create_label(data = get_eq_data() )
}
}
