% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location name of NOAA earthquake data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{a dataframe with raw data from the significant earthquake database}
}
\value{
returns a dataframe where the contents of the column \code{LOCATION_NAME} have been stripped of the country name
}
\description{
The NOAA significant earthquake data has some excess information in the column \code{LOCATION_NAME} concerning the country of the earthquake.
This function removes the country from the location name leaving only the location itself.
}
\note{
In some rare cases the \code{LOCATION_NAME} column has more than one country and location.
This function will keep only the last location that is present in the column. It removes all information before the last colon (:).
}
\examples{
\dontrun{

clean_data <- eq_clean_data(get_eq_data())
clean_data_location <- eq_location_clean(clean_data)
}

}
