% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_earthquakes.R
\name{eq_map}
\alias{eq_map}
\title{Create Leaflet Earthquake map}
\usage{
eq_map(data, annot_col = NA)
}
\arguments{
\item{data}{A (filtered and) cleaned dataframe with locations from the NOAA Earthquake database}

\item{annot_col}{A character string with the column that should be used for the pop_ups on the map}
}
\value{
A Leaflet map widget with the plotted locations of Earthquakes
}
\description{
This function creates a Leaflet map with the locations of earthquakes from the NOAA Earthquake database
}
\examples{
get_eq_data() \%>\%
 eq_clean_data() \%>\%
 eq_map(annot_col = "date")

}
