% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{A labeler geom for the geom_timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = NA, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to \code{ggplot()}.

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See \code{fortify()} for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a \code{data.frame}, and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? \code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. \code{borders()}.}

\item{n_max}{optional, a number to label only those earthquakes with the largest magnitudes. An aesthetic \code{magnitude} must be provided.}

\item{...}{other arguments passed on to \code{layer()}. These are often aesthetics, used to set an aesthetic to a fixed value, like \code{color = "red"} or \code{size = 3}. They may also be parameters to the paired geom/stat.}
}
\value{
Returns a layer that can be added to a ggplot object.
}
\description{
This geom creates labels to annotate the timeline(s) created by the \code{geom_timeline}.
}
\section{Aesthetics}{
 \code{geom_timeline} understands the following aesthetics :
\itemize{
\item{x, required, should be a date}
\item{label, required, the text to display}
\item{y, the same as in geom_timeline}
\item{magnitude, should be provided when n_max is used}
}
}

\examples{
\dontrun{

data <- get_eq_data() \%>\% eq_clean_data() \%>\% eq_location_clean()

ggplot(data, aes(x = date)) + geom_timeline() + geom_timeline_label(aes(label = LOCATION))

ggplot(data, aes(x = date)) +
    geom_timeline() +
    geom_timeline_label(aes(label = LOCATION, magnitude = EQ_PRIMARY), n_max = 5)
}
}
