% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_eq_data}
\alias{get_eq_data}
\title{Get significant earthquake data}
\usage{
get_eq_data()
}
\value{
Returns a dataframe with earthquake data
}
\description{
Retrieve the significant earthquake database from the NOAA website.

The Significant Earthquake Database contains information on destructive earthquakes from 2150 B.C. to the present
that meet at least one of the following criteria:
\itemize{
\item moderate damage, approximately $1 million or more
\item 10 or more deaths
\item magnitude 7.5 or greater
\item modified Mercalli Intensity X or greater
\item the earthquake generated a tsunami.
}

The database should be reference to as: National Geophysical Data Center / World Data Service (NGDC/WDS): Significant Earthquake Database. National Geophysical Data Center, NOAA. doi:10.7289/V5TD9V7K

More information can be found at the \href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{NOAA website}
}
\examples{
\dontrun{

data <- get_eq_data()
}
}
